requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
distlib==0.3.6
Django==4.1.3
django-phonenumber-field==7.0.2
Faker==15.3.4
filelock==3.8.2
h11==0.14.0
idna==3.4
outcome==1.2.0
phonenumbers==8.13.4
Pillow==9.4.0
platformdirs==2.5.4
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.13
virtualenv==20.17.1
wsproto==1.2.0
